+++ lib/src/constants.dart
@@ -1,8 +1,10 @@
 // ignore_for_file: constant_identifier_names
+
 part of 'bindings.dart';
 
 // Constants, https://github.com/zeromq/libzmq/blob/9bb6b2142385b78d47811b73745b9c21ec366106/include/zmq.h#L285-L301
 
+// Socket types
 const int ZMQ_PAIR = 0;
 const int ZMQ_PUB = 1;
 const int ZMQ_SUB = 2;
@@ -16,11 +18,13 @@ const int ZMQ_XPUB = 9;
 const int ZMQ_XSUB = 10;
 const int ZMQ_STREAM = 11;
 
+// Poller
 const int ZMQ_POLLIN = 1;
 const int ZMQ_POLLOUT = 2;
 const int ZMQ_POLLERR = 4;
 const int ZMQ_POLLPRI = 8;
 
+// Send/recv Options
 const int ZMQ_DONTWAIT = 1;
 const int ZMQ_SNDMORE = 2;
 
@@ -69,27 +73,72 @@ const int ZMQ_REQ_CORRELATE = 52;
 const int ZMQ_REQ_RELAXED = 53;
 const int ZMQ_CONFLATE = 54;
 const int ZMQ_ZAP_DOMAIN = 55;
+const int ZMQ_ROUTER_HANDOVER = 56;
+const int ZMQ_TOS = 57;
+const int ZMQ_CONNECT_ROUTING_ID = 61;
+const int ZMQ_GSSAPI_SERVER = 62;
+const int ZMQ_GSSAPI_PRINCIPAL = 63;
+const int ZMQ_GSSAPI_SERVICE_PRINCIPAL = 64;
+const int ZMQ_GSSAPI_PLAINTEXT = 65;
+const int ZMQ_HANDSHAKE_IVL = 66;
+const int ZMQ_SOCKS_PROXY = 68;
+const int ZMQ_XPUB_NODROP = 69;
+const int ZMQ_BLOCKY = 70;
+const int ZMQ_XPUB_MANUAL = 71;
+const int ZMQ_XPUB_WELCOME_MSG = 72;
+const int ZMQ_STREAM_NOTIFY = 73;
+const int ZMQ_INVERT_MATCHING = 74;
+const int ZMQ_HEARTBEAT_IVL = 75;
+const int ZMQ_HEARTBEAT_TTL = 76;
+const int ZMQ_HEARTBEAT_TIMEOUT = 77;
+const int ZMQ_XPUB_VERBOSER = 78;
+const int ZMQ_CONNECT_TIMEOUT = 79;
+const int ZMQ_TCP_MAXRT = 80;
+const int ZMQ_THREAD_SAFE = 81;
+const int ZMQ_MULTICAST_MAXTPDU = 84;
+const int ZMQ_VMCI_BUFFER_SIZE = 85;
+const int ZMQ_VMCI_BUFFER_MIN_SIZE = 86;
+const int ZMQ_VMCI_BUFFER_MAX_SIZE = 87;
+const int ZMQ_VMCI_CONNECT_TIMEOUT = 88;
+const int ZMQ_USE_FD = 89;
+const int ZMQ_GSSAPI_PRINCIPAL_NAMETYPE = 90;
+const int ZMQ_GSSAPI_SERVICE_PRINCIPAL_NAMETYPE = 91;
+const int ZMQ_BINDTODEVICE = 92;
 
+// 0MQ errors
 // https://github.com/zeromq/libzmq/blob/9bb6b2142385b78d47811b73745b9c21ec366106/include/zmq.h#L133
-const int _errorBase = 156384712;
+const int ZMQ_HAUSNUMERO = 156384712;
 
-const int ENOTSUP = _errorBase + 1;
-const int EPROTONOSUPPORT = _errorBase + 2;
-const int ENOBUFS = _errorBase + 3;
-const int ENETDOWN = _errorBase + 4;
-const int EADDRINUSE = _errorBase + 5;
-const int EADDRNOTAVAIL = _errorBase + 6;
-const int ECONNREFUSED = _errorBase + 7;
-const int EINPROGRESS = _errorBase + 8;
-const int ENOTSOCK = _errorBase + 9;
-const int EMSGSIZE = _errorBase + 10;
-const int EAFNOSUPPORT = _errorBase + 11;
-const int ENETUNREACH = _errorBase + 12;
-const int ECONNABORTED = _errorBase + 13;
-const int ECONNRESET = _errorBase + 14;
-const int ENOTCONN = _errorBase + 15;
-const int ETIMEDOUT = _errorBase + 16;
-const int EHOSTUNREACH = _errorBase + 17;
-const int ENETRESET = _errorBase + 18;
+const int ENOTSUP = ZMQ_HAUSNUMERO + 1;
+const int EPROTONOSUPPORT = ZMQ_HAUSNUMERO + 2;
+const int ENOBUFS = ZMQ_HAUSNUMERO + 3;
+const int ENETDOWN = ZMQ_HAUSNUMERO + 4;
+const int EADDRINUSE = ZMQ_HAUSNUMERO + 5;
+const int EADDRNOTAVAIL = ZMQ_HAUSNUMERO + 6;
+const int ECONNREFUSED = ZMQ_HAUSNUMERO + 7;
+const int EINPROGRESS = ZMQ_HAUSNUMERO + 8;
+const int ENOTSOCK = ZMQ_HAUSNUMERO + 9;
+const int EMSGSIZE = ZMQ_HAUSNUMERO + 10;
+const int EAFNOSUPPORT = ZMQ_HAUSNUMERO + 11;
+const int ENETUNREACH = ZMQ_HAUSNUMERO + 12;
+const int ECONNABORTED = ZMQ_HAUSNUMERO + 13;
+const int ECONNRESET = ZMQ_HAUSNUMERO + 14;
+const int ENOTCONN = ZMQ_HAUSNUMERO + 15;
+const int ETIMEDOUT = ZMQ_HAUSNUMERO + 16;
+const int EHOSTUNREACH = ZMQ_HAUSNUMERO + 17;
+const int ENETRESET = ZMQ_HAUSNUMERO + 18;
 
+// Native 0MQ error codes
+const int EFSM = ZMQ_HAUSNUMERO + 51;
+const int ENOCOMPATPROTO = ZMQ_HAUSNUMERO + 52;
+const int ETERM = ZMQ_HAUSNUMERO + 53;
+const int EMTHREAD = ZMQ_HAUSNUMERO + 54;
+
+// errno
+const int EINTR = 4;
+const int EBADF = 9;
 const int EAGAIN = 11;
+const int EACCES = 13;
+const int EFAULT = 14;
+const int EINVAL = 22;
+const int EMFILE = 24;