+++ lib/src/zeromq.dart
@@ -2,6 +2,7 @@ library dartzmq;
 
 import 'dart:async';
 import 'dart:collection';
+import 'dart:developer';
 import 'dart:ffi';
 import 'dart:io';
 import 'dart:typed_data';
@@ -10,38 +11,57 @@ import 'package:ffi/ffi.dart';
 
 import 'bindings.dart';
 
-String _platformPath(String name, {String? path}) {
-  path = path ?? "";
+String _platformPath(final String name, {String? path}) {
+  path = path ?? '';
   if (Platform.isLinux || Platform.isAndroid) {
-    return path + "lib" + name + ".so";
+    return path + 'lib' + name + '.so';
   }
   if (Platform.isMacOS) {
-    return path + "lib" + name + ".dylib";
+    return path + 'lib' + name + '.dylib';
   }
   if (Platform.isWindows) {
-    return path + name + ".dll";
+    return path + name + '.dll';
   }
-  throw Exception("Platform not implemented");
+  throw Exception('Platform not implemented');
 }
 
-DynamicLibrary _dlOpenPlatformSpecific(String name, {String? path}) {
+DynamicLibrary _dlOpenPlatformSpecific(final String name,
+    {final String? path}) {
   String fullPath = _platformPath(name, path: path);
   return DynamicLibrary.open(fullPath);
 }
 
 /// High-level wrapper around the ØMQ C++ api.
 class ZContext {
+  /// Native bindings
   late final ZMQBindings _bindings;
+
+  /// Native context
   late final ZMQContext _context;
+
+  /// Native poller
   late final ZMQPoller _poller;
 
+  /// Do we need to shutdown?
   bool _shutdown = false;
-  Timer? _timer;
 
-  final Map<ZMQSocket, ZSocket> _createdSockets = {};
-  final List<ZSocket> _listening = [];
+  /// Used for shutting down asynchronously
   Completer? _stopCompleter;
 
+  /// Timer used for running [_poll] function
+  Timer? _timer;
+
+  /// Keeps track of all sockets created by [createSocket].
+  /// Maps raw zeromq sockets [ZMQSocket] to our wrapper class [ZSocket].
+  final Map<ZMQSocket, ZSocket> _createdSockets = {};
+
+  /// Keeps track of all sockets that are currently being listened to
+  final List<ZSocket> _listening = [];
+
+  /// Create a new global ZContext
+  ///
+  /// Note only one context should exist throughout your application
+  /// and it should be closed if the app is disposed
   ZContext() {
     _initBindings();
     _context = _bindings.zmq_ctx_new();
@@ -50,29 +70,44 @@ class ZContext {
   }
 
   void _initBindings() {
-    // try {
-    _bindings = ZMQBindings(_dlOpenPlatformSpecific('libzmq-v142-mt-4_3_5'));
-    // } catch (err) {
-    //   log('Error loading bindings: ' + err.toString());
-    // }
+    final loaded = _loadBinding('zmq') || _loadBinding('libzmq-v142-mt-4_3_5');
+    if (!loaded) {
+      throw Exception('Could not load any zeromq library');
+    }
   }
 
+  bool _loadBinding(final String name) {
+    try {
+      _bindings = ZMQBindings(_dlOpenPlatformSpecific(name));
+      return true;
+    } catch (err) {
+      log('Failed to load library $name:  ${err.toString()}', name: 'dartzmq');
+    }
+    return false;
+  }
+
+  /// Shutdown zeromq. Will stop [_poll] asynchronously.
+  /// The returned [Future] will complete once [_poll] has been stopped
   Future stop() {
-    _shutdown = true;
     _stopCompleter = Completer();
+    _shutdown = true;
     return _stopCompleter!.future;
   }
 
+  /// Allocates memory and casts it to a [ZMQMessage]
   ZMQMessage _allocateMessage() {
     return malloc.allocate<Uint8>(64).cast();
   }
 
+  /// Starts the periodic polling task if it was not started already and
+  /// if there are actually listeners on sockets
   void _startPolling() {
     if (_timer == null && _listening.isNotEmpty) {
       _timer = Timer.periodic(const Duration(seconds: 1), (timer) => _poll());
     }
   }
 
+  /// Polling task receiving and handling socket messages
   void _poll() {
     final socketCount = _listening.length;
 
@@ -121,21 +156,21 @@ class ZContext {
 
     malloc.free(pollerEvents);
 
-    // After the polling iteration, re-schedule another one if necessary.
+    // Do we need to shutdown?
     if (_shutdown) {
       _shutdownInternal();
-      _stopCompleter?.complete(null);
     } else if (socketCount > 0) {
       return;
     }
 
-    // no polling necessary, reset flag so that the next call to _startPolling
-    // will bring the mechanism back up.
+    // If we land here either there are no
     _timer?.cancel();
     _timer = null;
+    _stopCompleter?.complete(null);
   }
 
-  ZSocket createSocket(SocketMode mode) {
+  /// Create a new socket of the given [mode]
+  ZSocket createSocket(SocketType mode) {
     final socket = _bindings.zmq_socket(_context, mode.index);
     final apiSocket = ZSocket(socket, this);
     _createdSockets[socket] = apiSocket;
@@ -171,10 +206,10 @@ class ZContext {
 
     _bindings.zmq_ctx_term(_context);
 
-    // final pollerPtrPtr = malloc.allocate<ZMQPoller>(0);
-    // pollerPtrPtr.value = _poller;
-    _bindings.zmq_poller_destroy(_poller);
-    malloc.free(_poller);
+    Pointer<ZMQPoller> pollerPointer = malloc.allocate<ZMQPoller>(0);
+    pollerPointer.value = _poller;
+    _bindings.zmq_poller_destroy(pollerPointer);
+    malloc.free(pollerPointer);
   }
 
   void _checkReturnCode(int code, {List<int> ignore = const []}) {
@@ -191,41 +226,62 @@ class ZContext {
   }
 }
 
-enum SocketMode {
+/// All types of sockets
+enum SocketType {
+  /// [ZMQ_PAIR] = 0
   pair,
+
+  /// [ZMQ_PUB] = 1
   pub,
+
+  /// [ZMQ_SUB] = 2
   sub,
+
+  /// [ZMQ_REQ] = 3
   req,
+
+  /// [ZMQ_REP] = 4
   rep,
+
+  /// [ZMQ_DEALER] = 5
   dealer,
+
+  /// [ZMQ_ROUTER] = 6
   router,
+
+  /// [ZMQ_PULL] = 7
   pull,
+
+  /// [ZMQ_PUSH] = 8
   push,
+
+  /// [ZMQ_XPUB] = 9
   xPub,
+
+  /// [ZMQ_XSUB] = 10
   xSub,
+
+  /// [ZMQ_STREAM] = 11
   stream
 }
 
-/// ZFrame
-///
-/// A 'frame' corresponds to one underlying zmq_msg_t in the libzmq code.
+/// A ZFrame or 'frame' corresponds to one underlying zmq_msg_t in the libzmq code.
 /// When you read a frame from a socket, the [hasMore] member indicates
-/// if the frame is part of an unfinished multipart message.
+/// if the frame is part of an unfinished multi-part message.
 class ZFrame {
   /// The payload that was received or is to be sent
   final Uint8List payload;
 
-  /// Is this frame part of an unfinished multipart message?
+  /// Is this frame part of an unfinished multi-part message?
   final bool hasMore;
 
   ZFrame(this.payload, {this.hasMore = false});
 }
 
-/// ZMessage
-///
-/// This class provides a list-like container interface,
+/// ZMessage provides a list-like container interface,
 /// with methods to work with the overall container. ZMessage messages are
 /// composed of zero or more ZFrame objects.
+// typedef ZMessage = Queue<ZFrame>;
 class ZMessage implements Queue<ZFrame> {
   final DoubleLinkedQueue<ZFrame> _frames = DoubleLinkedQueue();
 
@@ -310,7 +366,7 @@ class ZMessage implements Queue<ZFrame> {
   bool get isNotEmpty => _frames.isNotEmpty;
 
   @override
-  String join([String separator = ""]) => _frames.join(separator);
+  String join([String separator = '']) => _frames.join(separator);
 
   @override
   ZFrame get last => _frames.last;
@@ -367,17 +423,38 @@ class ZMessage implements Queue<ZFrame> {
   Iterable<T> whereType<T>() => _frames.whereType<T>();
 }
 
+/// ZeroMQ sockets present an abstraction of an asynchronous message queue,
+/// with the exact queuing semantics depending on the socket type in use.
+/// Where conventional sockets transfer streams of bytes or discrete datagrams,
+/// ZeroMQ sockets transfer discrete messages.
+///
+/// ZeroMQ sockets being asynchronous means that the timings of the physical
+/// connection setup and tear down, reconnect and effective delivery are
+/// transparent to the user and organized by ZeroMQ itself. Further, messages
+/// may be queued in the event that a peer is unavailable to receive them.
 class ZSocket {
+  /// Native socket
   final ZMQSocket _socket;
+
+  /// Global context
   final ZContext _context;
 
+  /// Has this socket been closed?
   bool _closed = false;
 
+  /// StreamController for providing received [ZMessage]s as a stream
   late final StreamController<ZMessage> _controller;
+
+  /// Stream of received [ZMessage]s
   Stream<ZMessage> get messages => _controller.stream;
+
+  /// Stream of received [ZFrame]s (expanded [messages] stream)
   Stream<ZFrame> get frames => messages.expand((element) => element._frames);
+
+  /// Stream of received payloads (expanded [frames] strem)
   Stream<Uint8List> get payloads => frames.map((e) => e.payload);
 
+  /// Construct a new [ZSocket] with a given underlying ZMQSocket [_socket] and the global ZContext [_context]
   ZSocket(this._socket, this._context) {
     _controller = StreamController(onListen: () {
       _context._listen(this);
@@ -386,12 +463,12 @@ class ZSocket {
     });
   }
 
-  /// Sends the [data] payload over this socket.
+  /// Sends the given [data] payload over this socket.
   ///
   /// The [more] parameter (defaults to false) signals that this is a multi-part
   /// message. ØMQ ensures atomic delivery of messages: peers shall receive
   /// either all message parts of a message or none at all.
-  void send(List<int> data, {bool more = false}) {
+  void send(final List<int> data, {final bool more = false}) {
     _checkNotClosed();
     final ptr = malloc.allocate<Uint8>(data.length);
     ptr.asTypedList(data.length).setAll(0, data);
@@ -403,7 +480,31 @@ class ZSocket {
     malloc.free(ptr);
   }
 
-  void bind(String address) {
+  /// Sends the given [frame] over this socket
+  ///
+  /// This is a convenience function and is the same as calling
+  /// [send(frame.payload, more: frame.hasMore)]
+  void sendFrame(final ZFrame frame) {
+    send(frame.payload, more: frame.hasMore);
+  }
+
+  /// Sends the given multi-part [message] over this socket
+  ///
+  /// This is a convenience function.
+  /// Note that the individual [ZFrame.hasMore] are ignored
+  void sendMessage(final ZMessage message) {
+    final lastIndex = message.length - 1;
+    for (int i = 0; i < message.length; ++i) {
+      send(message.elementAt(i).payload, more: i < lastIndex ? true : false);
+    }
+  }
+
+  /// Creates an endpoint for accepting connections and binds to it.
+  ///
+  /// The [address] argument is a string consisting of two parts as follows: 'transport://address'.
+  /// The transport part specifies the underlying transport protocol to use.
+  /// The meaning of the address part is specific to the underlying transport protocol selected.
+  void bind(final String address) {
     _checkNotClosed();
     final endpointPointer = address.toNativeUtf8();
     final result = _context._bindings.zmq_bind(_socket, endpointPointer);
@@ -411,7 +512,12 @@ class ZSocket {
     malloc.free(endpointPointer);
   }
 
-  void connect(String address) {
+  /// Connects the socket to an endpoint and then accepts incoming connections on that endpoint.
+  ///
+  /// The [address] argument is a string consisting of two parts as follows: 'transport://address'.
+  /// The transport part specifies the underlying transport protocol to use.
+  /// The meaning of the address part is specific to the underlying transport protocol selected.
+  void connect(final String address) {
     _checkNotClosed();
     final endpointPointer = address.toNativeUtf8();
     final result = _context._bindings.zmq_connect(_socket, endpointPointer);
@@ -419,6 +525,8 @@ class ZSocket {
     malloc.free(endpointPointer);
   }
 
+  /// Closes the socket and releases underlying resources.
+  /// Note after closing a socket it can't be reopened/reconncted again
   void close() {
     if (!_closed) {
       _context._handleSocketClosed(this);
@@ -428,21 +536,33 @@ class ZSocket {
     }
   }
 
-  void setOption(int option, String value) {
+  /// Set a socket [option] to a specific [value]
+  void setOption(final int option, final String value) {
     final ptr = value.toNativeUtf8();
     _context._bindings
         .zmq_setsockopt(_socket, option, ptr.cast<Uint8>(), ptr.length);
     malloc.free(ptr);
   }
 
-  void setCurvePublicKey(final String key) {
-    setOption(ZMQ_CURVE_PUBLICKEY, key);
-  }
-
+  /// Sets the socket's long term secret key.
+  /// You must set this on both CURVE client and server sockets, see zmq_curve(7).
+  /// You can provide the [key] as a 40-character string encoded in the Z85 encoding format.
   void setCurveSecretKey(final String key) {
     setOption(ZMQ_CURVE_SECRETKEY, key);
   }
 
+  /// Sets the socket's long term public key.
+  /// You must set this on CURVE client sockets, see zmq_curve(7).
+  /// You can provide the [key] as a 40-character string encoded in the Z85 encoding format.
+  /// The public key must always be used with the matching secret key.
+  void setCurvePublicKey(final String key) {
+    setOption(ZMQ_CURVE_PUBLICKEY, key);
+  }
+
+  /// Sets the socket's long term server key.
+  /// You must set this on CURVE client sockets, see zmq_curve(7).
+  /// You can provide the [key] as a 40-character string encoded in the Z85 encoding format.
+  /// This key must have been generated together with the server's secret key.
   void setCurveServerKey(final String key) {
     setOption(ZMQ_CURVE_SERVERKEY, key);
   }
@@ -470,11 +590,12 @@ class ZSocket {
 
   void _checkNotClosed() {
     if (_closed) {
-      throw StateError("This operation can't be performed on a cosed socket!");
+      throw StateError("This operation can't be performed on a closed socket!");
     }
   }
 }
 
+/// Custom Exception type for ZeroMQ specific exceptions
 class ZeroMQException implements Exception {
   final int errorCode;
 
@@ -490,8 +611,30 @@ class ZeroMQException implements Exception {
     }
   }
 
+  /// Maps error codes to messages
   static const Map<int, String> _errorMessages = {
-    EPROTONOSUPPORT: 'The requested transport protocol is not supported',
-    EADDRINUSE: 'The requested address is already in use',
+    // errno
+    EINTR: 'The operation was interrupted',
+    EBADF: 'Bad file descriptor',
+    // EAGAIN: '', // denpendant on what function has been called before
+    EACCES: 'Permission denied',
+    // EFAULT: '', // denpendant on what function has been called before
+    // EINVAL: '', // denpendant on what function has been called before
+    // EMFILE: '', // denpendant on what function has been called before
+
+    // 0MQ errors
+    ENOTSUP: 'Not supported',
+    EPROTONOSUPPORT: 'Protocol not supported',
+    ENOBUFS: 'No buffer space available',
+    ENETDOWN: 'Network is down',
+    EADDRINUSE: 'Address in use',
+    EADDRNOTAVAIL: 'Address not available',
+    ECONNREFUSED: 'Connection refused',
+    EINPROGRESS: 'Operation in progress',
+    EFSM: 'Operation cannot be accomplished in current state',
+    ENOCOMPATPROTO: 'The protocol is not compatible with the socket type',
+    ETERM: 'Context was terminated',
+    EMTHREAD: 'No thread available',
+    EHOSTUNREACH: 'Host unreachable',
   };
 }